+++ src/app/layout.tsx
@@ -27,11 +27,11 @@ export default function RootLayout({ children }: Readonly<{ children: React.Reac
   return (
     <html lang="en">
       <body>
+        <Analytics />
         <Header />
         {children}
         <DisclaimerFooter />
         <BottomNav />
-        <Analytics />
       </body>
     </html>
   );